+++ src/modules/eventHandlers.js
@@ -1,6 +1,5 @@
-import TodoList from './taskList';
 import ProjectList from './projectList'
-
+/*
 export default function eventHandlers() {
     const todoList = new TodoList()
     const projectList = new ProjectList()
@@ -77,3 +76,4 @@ export default function eventHandlers() {
         });
     });
 }
+*/
