+++ src/modules/project.js
@@ -1,8 +1,30 @@
-import TodoList from './taskList.js'
+import Task from './task'
 
-export default class Project extends TodoList {
+export default class Project {
     constructor(name) {
         this.name = name;
-        // perhaps todolist should be a child of the project class, inheriting the task objects from the project. idrk.
+        this.tasks = [];
+        this.tasks.push(new Task('Vacuum the floor', 'unchecked'));
+        this.tasks.push(new Task('Do the dishes', 'unchecked'));
+        this.tasks.push(new Task('Fold the washing', 'unchecked'));
+    }
+
+    setTasks(tasks) {
+        this.tasks = tasks;
+    }
+
+    getTasks() {
+        return this.tasks;
+    }
+
+    addTask(taskName, status) {
+        this.tasks.push(new Task(taskName, status));
+    }
+
+    deleteTask(taskName) {
+        const index = this.tasks.indexOf(taskName);
+        if (index > -1) {
+            this.tasks.splice(index, 1);
+        }
     }
 }
